This is related to #4.
CC: @TatriX