@TatriX 2016-05-11
I'm working on the `kite-mini` package and I want to add sourcemap support to the repl.
When chrome debugger tells me that he parsed a script he also gives me a sourcemap url.
When I'm getting a `Console.added` message I'd like to use mapped location so I need a parsed sourcemap (`sourcemap-original-position-for` actually) 

I'm not sure if I should lazy load sourcemaps or `emacs-sourcemap` should lazy parse it's mappings. What do you think?