@TatriX 2016-05-11
I guess sourcemap shoudn't lazy parse mappings so user will get long creation of the sourcemap and not long first call to the `sourcemap-original-position-for` etc.