@TatriX 2016-05-11
I've tested this branch and implemented lazy load for `kite-mini`. Parsing of the large sourcemaps (more than 40000 lines) is dead slow. These big sourcemap generated for lib bundles so I hope I could manage it on my side.

Thank you!