@syohex 2016-05-12
```lisp
(defvar test-input
"{
  \"version\": 3,
  \"file\": \"min.js\",
  \"names\": [\"bar\", \"baz\", \"n\"],
  \"sources\": [\"one.js\", \"two.js\"],
  \"sourceRoot\": \"http://example.com/www/js/\",
  \"mappings\": \"CAAC,IAAI,IAAM,SAAUA,GAClB,OAAOC,IAAID;CCDb,IAAI,IAAM,SAAUE,GAClB,OAAOA\"
}")

(setq sourcemap (sourcemap-from-string test-input))
(benchmark 1000 '(sourcemap-original-position-for sourcemap :line 2 :column 28))
```

Micro benchmark between original version and optimize version(Change data structure and use binary search instead of linear search)

```
Original: Elapsed time: 0.985985s (0.510061s in 65 GCs)
Optimized: Elapsed time: 0.097774s (0.043636s in 3 GCs) ;; 10 times faster
```