@syohex 2016-05-12
Emacs Lisp is slow language. You should write/use sourcemap decoder written in fast language if you need more speed.